requirements.txt
torch>=1.8.0
numpy>=1.19.2
biopython>=1.79
scikit-learn>=0.24.2
tqdm>=4.62.3
wandb>=0.12.0
psutil>=5.8.0
enformer-pytorch>=0.0.1 